qrCodeGenerator/GenerateQRCode/requirement.txt
azure-functions
azure-storage-blob
qrcode
pillow